refactor: ProgressStatus 열거형 Task 타입 내부로 이동
