--- ProjectManager/ProjectManager/Model/ProgressStatus.swift
@@ -1,18 +1,0 @@
-import Foundation
-
-enum ProgressStatus: String {
-    case todo
-    case doing
-    case done
-    
-    var name: String {
-        switch self {
-        case .todo:
-            return "Todo"
-        case .doing:
-            return "Doing"
-        case .done:
-            return "Done"
-        }
-    }
-}
