+++ ProjectManager/ProjectManager/View/TaskListCellView.swift
@@ -5,8 +5,8 @@ struct TaskListCellView: View {
     
     @State private var isShowTaskDetailView = false
     @State private var isShowUpdateTaskState = false
-    @State private var firstMoveStatus: ProgressStatus = .doing
-    @State private var secondMoveStatus: ProgressStatus = .done
+    @State private var firstMoveStatus: Task.ProgressStatus = .doing
+    @State private var secondMoveStatus: Task.ProgressStatus = .done
     
     var task: Task
